Separate settings list.
 https://github.com/Esri/arcgis-maps-sdk-swift-samples/pull/294#discussion_r1391622402
