+++ Shared/Samples/Find route around barriers/FindRouteAroundBarriersView.swift
@@ -25,9 +25,6 @@ struct FindRouteAroundBarriersView: View {
     /// A Boolean value indicating whether a routing operation is in progress.
     @State private var routingIsInProgress = false
     
-    /// A Boolean value indicating whether routing will find the best sequence.
-    @State private var routingFindsBestSequence = false
-    
     /// The geometry of a direction maneuver to set the viewpoint to.
     @State private var directionGeometry: Geometry?
     
@@ -80,6 +77,7 @@ struct FindRouteAroundBarriersView: View {
                             guard routingIsInProgress else { return }
                             
                             do {
+                                // Route when the button is pressed
                                 try await model.route()
                                 
                                 // Update the viewpoint to the geometry of the new route.
@@ -109,12 +107,12 @@ struct FindRouteAroundBarriersView: View {
                         } label: {
                             Image(systemName: "arrow.triangle.turn.up.right.diamond")
                         }
+                        .disabled(model.route == nil)
                         .task(id: directionGeometry) {
                             guard let directionGeometry else { return }
                             model.directionGraphic.geometry = directionGeometry
                             await mapViewProxy.setViewpointGeometry(directionGeometry, padding: 100)
                         }
-                        .disabled(model.route == nil)
                         Spacer()
                         
                         Picker("Features", selection: $featuresSelection) {
@@ -126,31 +124,8 @@ struct FindRouteAroundBarriersView: View {
                         Spacer()
                         
                         SheetButton(title: "Settings") {
-                            List {
-                                Toggle(isOn: $routingFindsBestSequence) {
-                                    Text("Find Best Sequence")
-                                }
-                                .onChange(of: routingFindsBestSequence) { newValue in
-                                    model.routeParameters.findsBestSequence = newValue
-                                }
-                                
-                                Section {
-                                    Toggle(isOn: Binding(
-                                        get: { model.routeParameters.preservesFirstStop },
-                                        set: { model.routeParameters.preservesFirstStop = $0 }
-                                    )) {
-                                        Text("Preserve First Stop")
-                                    }
-                                    
-                                    Toggle(isOn: Binding(
-                                        get: { model.routeParameters.preservesLastStop },
-                                        set: { model.routeParameters.preservesLastStop = $0 }
-                                    )) {
-                                        Text("Preserve Last Stop")
-                                    }
-                                }
-                                .disabled(!routingFindsBestSequence)
-                            }
+                            SettingsList()
+                                .environmentObject(model)
                         } label: {
                             Image(systemName: "gear")
                         }
@@ -178,69 +153,5 @@ struct FindRouteAroundBarriersView: View {
             }
         }
         .alert(isPresented: $errorAlertIsShowing, presentingError: error)
-    }
-}
-
-private extension FindRouteAroundBarriersView {
-    /// A button with a given label that brings up a sheet containing given content.
-    struct SheetButton<Content: View, Label: View>: View {
-        /// The string to display as the title of the sheet
-        let title: String
-        
-        /// A view that contains the content to display in the sheet.
-        @ViewBuilder let content: () -> Content
-        
-        /// A view that describes the purpose of the button.
-        @ViewBuilder let label: () -> Label
-        
-        /// A Boolean value indicating whether the sheet is showing.
-        @State private var sheetIsShowing = false
-        
-        var body: some View {
-            if #available(iOS 16, *) {
-                button
-                    .popover(isPresented: $sheetIsShowing, arrowEdge: .bottom) {
-                        sheetContent
-                            .presentationDetents([.fraction(0.5)])
-#if targetEnvironment(macCatalyst)
-                            .frame(minWidth: 300, minHeight: 270)
-#else
-                            .frame(minWidth: 320, minHeight: 390)
-#endif
-                    }
-            } else {
-                button
-                    .sheet(isPresented: $sheetIsShowing, detents: [.medium]) {
-                        sheetContent
-                    }
-            }
-        }
-        
-        /// The button that presents the sheet.
-        @ViewBuilder private var button: some View {
-            Button {
-                sheetIsShowing = true
-            } label: {
-                label()
-            }
-        }
-        
-        /// The content to display in the sheet with a title and done button.
-        @ViewBuilder private var sheetContent: some View {
-            NavigationView {
-                content()
-                    .navigationTitle(title)
-                    .navigationBarTitleDisplayMode(.inline)
-                    .toolbar {
-                        ToolbarItem(placement: .confirmationAction) {
-                            Button("Done") {
-                                sheetIsShowing = false
-                            }
-                        }
-                    }
-            }
-            .navigationViewStyle(.stack)
-            .frame(idealWidth: 320, idealHeight: 428)
-        }
     }
 }
